ai/requirements.txt
# AI servers
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.0.1

# langchain
langchain==0.3.25
langchain-core==0.3.58
langchain-openai==0.3.16
langchain-ollama==0.3.2
langchain-anthropic~=0.3.12

pydantic==2.11.4

# infrastructure
aiokafka==0.10.0
kafka-python==2.0.2
httpx==0.27.0
wheel==0.46.1
motor==3.7.0

# Test dependencies
pytest==8.3.5
pytest-asyncio==0.23.6
pytest-mock==3.12.0
pytest-cov==4.1.0
